+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.2] - 2020-02-25
+
+### Updated
+
+* Possibility to overwrite the URL from the monitoring service in a monitoring widget (allows to monitor one address by box widget).
+* Fix a small display bug (forgot to delete a useless debug output... oops).
+
+## [0.4.1] - 2020-11-24
+
+### Updated
+
+* Fix filepaths when using DevDash on Windows
+
 ## [0.4.0] - 2020-06-08
 
 ### Added